+++ lib/ui/component/unfocus.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class Unfocus extends StatelessWidget {
+  const Unfocus({
+    super.key,
+    required this.child,
+  });
+
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      behavior: HitTestBehavior.opaque,
+      onTap: () {
+        FocusManager.instance.primaryFocus?.unfocus();
+      },
+      child: child,
+    );
+  }
+}
